% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canton_CH.R
\name{canton_CH}
\alias{canton_CH}
\alias{canton_namesStrict}
\title{Swiss canton names - convert them to different languages and abbreviations}
\usage{
canton_CH()

canton_namesStrict(query, output = "iso2")
}
\arguments{
\item{query}{a character vector of Swiss canton names in any language}

\item{output}{a character string, the ouput format}
}
\description{
Functions to convert Switzerland's cantons, districts and communes names to their abbreviations and official BFS ID
}
\details{
a data.frame used by \code{canton_namesStrict}

See \code{canton_CH}
}
\examples{
\dontrun{
  canton_CH()
}
canton_query <- c('Argovie', 'Appenzell Rh.-I.', 'St. Gallen', 'Zürich', 'Vaud')
canton_namesStrict(canton_query)
canton_namesStrict(canton_query, output = "eng")
}
