% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commmune_suisse_api.R
\name{commune_suisse_api}
\alias{commune_suisse_api}
\title{Helper function to query OFS commune application}
\usage{
commune_suisse_api(name = "", start_time = "01.01.2019",
  end_time = "01.01.2019", minDist = 0.95)
}
\arguments{
\item{name, }{a character string the name of the muncipality to query}

\item{start_time, end_time}{a character string the start and end dates as dd.mm.yyy}

\item{minDist}{minimum Levenshtein distance for string matching}
}
\description{
Basic function to query ofs commune app
}
\examples{
\dontrun{
commune_suisse_api(name = "Morges")
}
}
\seealso{
https://www.agvchapp.bfs.admin.ch/fr/communes/query
}
