% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryTranslation.R
\docType{data}
\name{countryTranslation}
\alias{countryTranslation}
\alias{countryTranslations}
\title{country translation}
\format{An object of class \code{data.frame} with 281 rows and 13 columns.}
\usage{
countryTranslations

countryTranslation(query, output)
}
\arguments{
\item{query}{a character vector of iso2, iso3 or iso numeric country codes}

\item{output}{a character vector containing the iso2 output languages ("EN", "DE", "FR", "IT", "ES", "PT","RU", "ZH", "JA","AR")}
}
\value{
a data.frame of dimension \code{query} by \code{output}
}
\description{
Functions to convert country names given in iso2 or iso3 to 10 different languages

Convert country names to 10 possible languages
}
\examples{
print(countryTranslations)
countryTranslation(query = c('AF', 'FR', 'US', 'CH', 'JP'), output = "EN")
}
\keyword{datasets}

