% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communes_ch_indicators.R
\name{loadCHzipcode}
\alias{loadCHzipcode}
\title{Load the NPA/PLZ/ Zip code for Switzerland}
\usage{
loadCHzipcode()
}
\value{
a data.frame
}
\description{
Load the NPA/PLZ/ Zip code for Switzerland
}
\seealso{
\url{http://opendata.admin.ch/de/dataset/ch-swisstopo-vd-ortschaftenverzeichnis_plz/resource/35001b61-e7c1-4124-89fa-17fac7b1139e} from: \url{http://opendata.admin.ch/de/dataset/ch-swisstopo-vd-ortschaftenverzeichnis_plz}
}
