% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communes__ch_geographicalLevels.R
\name{loadCommunesCHgeographicalLevels}
\alias{loadCommunesCHgeographicalLevels}
\title{Swiss communes geographical data}
\usage{
loadCommunesCHgeographicalLevels()
}
\value{
a data.frame with tons of geographical features, check the source excel for more details
}
\description{
Load the Swiss statistical office communes data from: \url{https://www.bfs.admin.ch/bfs/fr/home/bases-statistiques/niveaux-geographiques.html}
}
\details{
The portrait includes non political OFS ID, i.e. lakes
}
\examples{
data <- loadCommunesCHgeographicalLevels()
head(data)
\dontrun{

}
}
\seealso{
Typologie des communes et typologie urbain-rural 2012: \url{https://www.bfs.admin.ch/bfs/fr/home/actualites/quoi-de-neuf.assetdetail.2543324.html}
}
