% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapShapefiles.R
\name{mapShapefiles}
\alias{formatShp}
\alias{getPathShp}
\alias{mapShapefiles}
\title{Geographical map with shapefiles}
\usage{
getPathShp(geo = "CH", year = 2016)

formatShp(shp)
}
\arguments{
\item{geo}{a character, either 'CH' or 'world'}

\item{year}{a numeric, year for data to use (default 2015)}

\item{shp}{shapefile contents loaded ojbect, as returned by \code{readOGR}}
}
\value{
a character, the path to the shapefiles folder

a dataframe object
}
\description{
Helper functions to draw geographical map based on shapefiles
}
\examples{
require(rgdal)
require(rgeos)
require(ggplot2)
require(maptools)
path <- getPathShp('world')
layers <-  ogrListLayers(path)
# read shapefiles for given layer
world <- readOGR(path, layer = layers[1])
world.df <- formatShp(world)
# plot world map
wm <- ggplot(world.df, aes(x = long, y = lat, group = group)) +
  geom_polygon(size = 0.01, aes(fill = FIPS)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid = element_blank(),
    axis.ticks = element_blank(),
    axis.title = element_blank(),
    axis.text = element_blank()
  )
 wm
 # With nice orthographic projection
 wm + coord_map("ortho")
# Same with swiss commune map (slow!)
\dontrun{
path.ch <- getPathShp('CH', year = 2015)
ch <- readOGR(path.ch, layer = 'municipalities-without-lakes')
ch.df <- formatShp(ch)
# plot swiss commune map
ch.communes <- ggplot(ch.df, aes(x = long, y = lat, group = group)) + geom_polygon(size = 0, aes(fill = GEMNAME)) +
theme_minimal() + theme(legend.position = "none", panel.grid = element_blank(), axis.ticks = element_blank(), axis.title = element_blank(), axis.text = element_blank())
ch.communes
# add the lake
lake.df <- formatShp(readOGR(path.ch, layer = "lakes"))
ch.communes + geom_polygon(data = lake.df, fill = "blue", colour = "blue")
# get all the id and name of communes from the map
bfsnName <- ch.df[!duplicated(ch.df$BFSNR),c('BFSNR', 'GEMNAME')]
bfsnName[order(bfsnName[,1]),]
}
}
\seealso{
https://github.com/hadley/ggplot2/wiki/plotting-polygon-shapefiles
}

