% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofs_px_translate.R
\name{ofsPx_wrangle}
\alias{ofsPx_wrangle}
\title{Get the French translation from an OFS typical px data file}
\usage{
ofsPx_wrangle(px.file, langout = "fr", attachCode = "", ...)
}
\arguments{
\item{px.file}{a string, the path to an OFS px file}

\item{langout}{a string, the output language. Has to be one of fr, it, en}

\item{attachCode}{a vector of strings, the column names (in the output language) for which to add the code identifiers}

\item{\dots}{Additional arguments passed to pxR::read.px(). encoding is especially useful}
}
\value{
a data.frame
}
\description{
A highlevel wrapper that translates the column names and 
values from German to French. It can also fetch OFS' internal 
code identifiers (useful for mapping). It assumes the provided data 
follows a well defined structure and that it is in German
}
\details{
BFS-STAT TAB px data cubes \url{https://www.pxweb.bfs.admin.ch/pxweb/fr/?rxid=e8ebe1b6-5756-49f0-8bfb-9eaa1cf758b2}
}
\examples{
\dontrun{
}
}
