% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantonal_ch_federalBallot.R
\name{processFederalBallotByCantons}
\alias{processFederalBallotByCantons}
\title{Process Portraits regionaux de la Suisse canton px file to be used by loadCantonsCHFederalBallot
This will download the px file from \url{https://www.bfs.admin.ch/bfs/en/home/statistics/politics/popular-votes.assetdetail.1363949.html}, 
process it and save it as a Rdata file in inst/extdata. 
Run it when new data are available.}
\arguments{
\item{url}{the URL to the px file with all federal ballots}

\item{output}{the output file name to be saved in the package inst/extdata}
}
\description{
Process Portraits regionaux de la Suisse canton px file to be used by loadCantonsCHFederalBallot
This will download the px file from \url{https://www.bfs.admin.ch/bfs/en/home/statistics/politics/popular-votes.assetdetail.1363949.html}, 
process it and save it as a Rdata file in inst/extdata. 
Run it when new data are available.
}
\examples{
\dontrun{
processPortraitsRegionauxCantons()
}
}
\seealso{
\url{https://www.bfs.admin.ch/bfs/en/home/statistics/politics/popular-votes.assetdetail.1363949.html}
}
