% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantonal_ch_indicators.R
\name{processPortraitsRegionauxCantons}
\alias{processPortraitsRegionauxCantons}
\title{Process Portraits regionaux de la Suisse canton xls}
\usage{
processPortraitsRegionauxCantons(
  file = "je-f-21.03.02.xlsx",
  output = "cantonal_CH_2020_indicators_je-f-21.03.02.csv"
)
}
\arguments{
\item{output}{the output file name to be saved in the package inst/extdata}
}
\description{
Run it when new data are available: \url{https://www.bfs.admin.ch/bfs/en/home/statistics/catalogues-databases.assetdetail.je-f-21.03.02.html}
Warning of possible break if the OFS changes the data structure of the xls file!!
}
\examples{
\dontrun{
tamMap:::processPortraitsRegionauxCantons()
}
}
\seealso{
\url{https://www.bfs.admin.ch/bfs/en/home/statistics/catalogues-databases.assetdetail.je-f-21.03.02.html}
}
