% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_cities.R
\name{processSwissCities}
\alias{processSwissCities}
\title{Main Swiss cities as simple feature object}
\usage{
processSwissCities(
  input = "T_STST_2018_01_1.xlsx",
  output = "swiss_cities.shp"
)
}
\arguments{
\item{input}{the xslx file name Population from the Swiss cities Union in inst/extdata}
}
\value{
NULL it will save
}
\description{
Script use to generate a shapefile with Swiss cities, their population and their LV95 coordinates. 
Its output needs to be saved in inst/extdata/CH/villes !
To be used by \code{shp_path}
}
\details{
From the \href{https://uniondesvilles.ch/fr/Info/Documentation/Statistiques_des_villes_suisses}{Union des villes suisses} the excel file (Chapitre population) was used a population source (31.12.2016)
}
