% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svgTranslate.R
\name{svgTranslate}
\alias{createTextToTranslate}
\alias{createTranslatedSVG}
\alias{svgTranslate}
\title{Collection of functions to get and replace text in a svg file}
\usage{
createTextToTranslate(input, ouputFileAppend = "_text.csv")

createTranslatedSVG(input = NULL, tradFile = NULL, inDirectory = "trad",
  overwrite = TRUE)
}
\arguments{
\item{input}{path to a svg file}

\item{ouputFileAppend}{suffix to append at the end of the created SVG file, if NULL returns a vector of text elements}

\item{tradFile}{path to the csv translations}

\item{inDirectory}{path to a directory where all the output files are saved}

\item{overwrite}{a boolean, output svg files to be overwritten?}

\item{input}{path to the original svg file to translate}
}
\value{
the output file path or a character vector if ouputFileAppend is NULL
}
\description{
createTextToTranslate: get the path to a svg file and return a csv file with all text elements (which are not pure numerics)

createTranslatedSVG: get a svg file and its text translations, output the translated svgs
}
\examples{
require(XML)
test <- createTextToTranslate(system.file("extdata", "slopegraph_test.svg", package="swiMap"), ouputFileAppend = NULL)
test
}

