% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canton_CH.R
\docType{data}
\name{swi_CHnames}
\alias{canton_CH}
\alias{canton_namesStrict}
\alias{swi_CHnames}
\title{Swiss names}
\format{An object of class \code{data.frame} with 26 rows and 22 columns.}
\usage{
canton_CH

canton_namesStrict(query, output = "iso2")
}
\arguments{
\item{query}{a character vector of Swiss canton names in any language}

\item{output}{a character string, the ouput format}
}
\description{
Functions to convert Switzerland's cantons, districts and communes names to their abbreviations and official BFS ID
TODO: expand countrycode for more languages (see in inst/ countrynames.csv )

Convert Swiss canton names to different languages and abbreviations
}
\details{
See \code{canton_CH}
}
\examples{
print(canton_CH)
canton_query <- c('Argovie', 'Appenzell Rh.-I.', 'St. Gallen', 'Zürich', 'randomChar', 'Vaudx')
canton_namesStrict(canton_query)
canton_namesStrict(canton_query, output = "english")
}
\keyword{datasets}

