% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_theme.R
\name{theme_map}
\alias{theme_map}
\title{ggplot2 bare theme for map}
\usage{
theme_map(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A basic ggplot2 theme to use for map without graticule/grid lines
}
\examples{
require(ggplot2)
ggplot() + 
  geom_sf(data = tilemap_ch()) + 
  theme_map()
}
