% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canton_CH.R
\docType{data}
\name{canton_CH}
\alias{canton_CH}
\title{Swiss canton names}
\format{An object of class \code{data.frame} with 26 rows and 23 columns.}
\usage{
canton_CH
}
\description{
Functions to convert Switzerland's cantons, districts and communes names to their abbreviations and official BFS ID
}
\examples{
\dontrun{
print(canton_CH)
}
}
\keyword{datasets}
