% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantonal_ch_federalBallot.R
\name{loadCantonsCHFederalBallot}
\alias{loadCantonsCHFederalBallot}
\alias{processFederalBallotByCantons}
\title{Federal ballot results by cantons}
\usage{
loadCantonsCHFederalBallot(file = "federalBallot_cantons.RData")


  processFederalBallotByCantons(url = "https://www.pxweb.bfs.admin.ch/DownloadFile.aspx?file=px-x-1703030000_100",
  output = "federalBallot_cantons.RData")
}
\arguments{
\item{file}{the name of the csv file processed by processFederalBallotByCantons to load}

\item{url}{the URL to the px file with all federal ballots}

\item{output}{the output file name to be saved in the package inst/extdata}
}
\value{
a matrix, rownames are cantons (2 letters name) and colnames are federal ballot IDs. Check the attributes ballotName and date
}
\description{
Load any federal ballots resuts by cantons: https://www.pxweb.bfs.admin.ch/Selection.aspx?px_language=fr&px_db=px-x-1703030000_100&px_tableid=px-x-1703030000_100\px-x-1703030000_100.px&px_type=PX

Process Portraits régionaux de la Suisse canton px file
This will download the px file from \url{https://www.bfs.admin.ch/bfs/en/home/statistics/politics/popular-votes.assetdetail.1363949.html}, process it and save it as a Rdata file
Necessary to run it before using loadCantonsCHFederalBallot
}
\details{
Run processFederalBallotByCantons() to generate a readable csv file for loadCantonsCHFederalBallot
}
\examples{
\dontrun{
processPortraitsRegionauxCantons()
}
}
\seealso{
\url{https://www.bfs.admin.ch/bfs/en/home/statistics/politics/popular-votes.assetdetail.1363949.html}
}

