% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/communes_ch_indicators.R
\name{communes_CH_indicators}
\alias{communes_CH_indicators}
\alias{loadCHzipcode}
\alias{loadCommunesCHportraits}
\title{Swiss communes socio-demographic indicators}
\usage{
loadCommunesCHportraits()

loadCHzipcode()
}
\value{
a matrix

a data.frame
}
\description{
Load socio-demographic data by the Swiss Statistical Office.

Load the NPA/PLZ/ Zip code for Switzerland
}
\details{
The matrix has the commune BFS code as rowname and 3 attributes: \itemize{
\item \code{communeName} the text name of the commune (same length as the matrix length)
\item \code{indicatorYear} & \code{indicatorGroup} respectively the year and the category of the communal indicator (both of same length as the matrix width). See \url{http://www.bfs.admin.ch/bfs/portal/fr/index/regionen/02/key.html}
}
}
\examples{
communeData <- loadCommunesCHportraits()
colnames(communeData)
rownames(communeData)

# Select only "surface" indicators
colIdx <- which(attr(communeData, "indicatorGroup") == "Surface")
head(communeData[,colIdx])

zipcode <- loadCHzipcode()
match(zipcode$Gemeindename, attr(communeData, "communeName"))
}
\seealso{
Office fédéral de la statistique > Les Régions > Communes > Données et explications (portraits): \url{http://www.bfs.admin.ch/bfs/portal/fr/index/regionen/02/daten.html} \cr
The cleaning up of raw xls file: \itemize{
  \item Remove all the empty lines (header has 3 lines: Indicator type, indicator, year)
  \item Reorder the header for: indicator type, year, indicator
  \item Remove in the data table, the line for Switzerland
  \item Remove the text at the bottom
  \item Save as csv
}

\url{http://opendata.admin.ch/de/dataset/ch-swisstopo-vd-ortschaftenverzeichnis_plz/resource/35001b61-e7c1-4124-89fa-17fac7b1139e} from: \url{http://opendata.admin.ch/de/dataset/ch-swisstopo-vd-ortschaftenverzeichnis_plz}
}

