% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countryTranslation.R
\docType{data}
\name{countryTranslation}
\alias{countryTranslation}
\alias{countryTranslations}
\title{country translation}
\format{\preformatted{'data.frame':	281 obs. of  13 variables:
 $ iso2      : chr  "AF" "AL" "DZ" "AO" ...
 $ iso3      : chr  "AFG" "ALB" "DZA" "AGO" ...
 $ isoNumeric: chr  "4" "8" "12" "24" ...
 $ EN        : chr  "Afghanistan" "Albania" "Algeria" "Angola" ...
 $ DE        : chr  "Afghanistan" "Albanien" "Algerien" "Angola" ...
 $ FR        : chr  "Afghanistan" "Albanie" "Algérie" "Angola" ...
 $ IT        : chr  "Afghanistan" "Albania" "Algeria" "Angola" ...
 $ ES        : chr  "Afganistán" "Albania" "Argelia" "Angola" ...
 $ PT        : chr  "Afeganistão" "Albânia" "Argélia" "Angola" ...
 $ RU        : chr  "Афганистан" "Албания" "Алжир" "Ангола" ...
 $ ZH        : chr  "阿富汗" "阿尔巴尼亚" "阿尔及利亚" "安哥拉" ...
 $ JA        : chr  "アフガニスタン" "アルバニア" "アルジェリア" "アンゴラ" ...
 $ AR        : chr  "أفغانستان" "ألبانيا" "الجزائر" "أنغولا" ...
}}
\usage{
countryTranslations

countryTranslation(query, output)
}
\arguments{
\item{query}{a character vector of iso2, iso3 or iso numeric country codes}

\item{output}{a character vector containing the iso2 output languages ("EN", "DE", "FR", "IT", "ES", "PT","RU", "ZH", "JA","AR")}
}
\value{
a data.frame of dimension \code{query} by \code{output}
}
\description{
Functions to convert country names given in iso2 or iso3 to 10 different languages

Convert country names to 10 possible languages
}
\examples{
print(countryTranslations)
countryTranslation(query = c('AF', 'FR', 'US', 'CH', 'JP'), output = "EN")
}
\keyword{datasets}

