% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communes__ch_geographicalLevels.R
\name{loadCommunesCHgeographicalLevels}
\alias{loadCommunesCHgeographicalLevels}
\title{Swiss communes geographical data}
\usage{
loadCommunesCHgeographicalLevels()
}
\value{
a data.frame with tons of geographical features, check the source excel for more details
}
\description{
Load the Swiss statistical office communes data from: \url{http://www.bfs.admin.ch/bfs/portal/fr/index/regionen/thematische_karten/01/01.html}
}
\examples{
data <- loadCommunesCHgeographicalLevels()
head(data)
}

