% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLimitesGeneralises_OFS.R
\name{shp_path}
\alias{shp_path}
\title{Return path to OFS geospatial data}
\usage{
shp_path(y = 2018, features = c("municipalities", "cantons", "lakes",
  "country"), dirGeo = "CH")
}
\arguments{
\item{y, }{a numeric of lenghth 1. The year (as of the 1st of Jan) of geo data to get}

\item{features, }{a string vector with the geographical levels' paths to returns, one of municipalities, municipalities_encl, lakes, agglomerations, cantons, largeRegions and country}
}
\value{
a named vector with the full path to the shapefiles, name the geographical levels
}
\description{
NOTE: Should be deprecated as soon the same data is available from swisstopo API!!
Manual work, download all the G2 resolutions shp (non-vz) locally. 
So far 2018 until 2013 downloaded and processed.
Use that script to rename all the downloaded file in a folder and move them to:
inst/extdata/shp/CH/
}
\examples{
shp_ch_paths_2018 <- shp_path(2018)
\dontrun{
require(tidyverse)
require(sf)
# loop and load the geo data in a named list
shp_ch_geodata <- shp_ch_paths_2018 \%>\% map(function(x) {
  layerName <- st_layers(x)
  st_read(x, layer = layerName$name) 
  }
)

theme_map <- theme(
legend.position = "bottom", 
panel.background = element_rect(fill = "transparent"),
panel.grid = element_line(colour = 'transparent'),
axis.ticks = element_blank(), axis.title = element_blank(),
axis.text = element_blank()
)

# plot
gp <- ggplot() +
  geom_sf(data = shp_ch_geodata$municipalities, aes(fill = AREA_HA), lwd = 0.05, colour = "#0d0d0d") +
  geom_sf(data = shp_ch_geodata$cantons, lwd = 0.15, colour = "#333333", fill = NA) +
  geom_sf(data = shp_ch_geodata$country, lwd = 0.25, colour = "#000d1a", fill = NA) +
  geom_sf(data = shp_ch_geodata$lakes, lwd = 0, fill = "#0066cc")
  gp + 
  coord_sf(datum = NA, expand = F) +
  theme_map
}
}
\seealso{
\url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/limites-administratives/limites-communales-generalisees.html}
}
