% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/canton_CH.R
\docType{data}
\name{swi_CHnames}
\alias{canton_CH}
\alias{canton_namesStrict}
\alias{swi_CHnames}
\title{Swiss names}
\format{\preformatted{'data.frame':	26 obs. of  16 variables:
 $ iso2     : Factor w/ 26 levels "AG","AI","AR",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ iso4     : Factor w/ 26 levels "CH-AG","CH-AI",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ français : Factor w/ 26 levels "Appenzell Rhodes-Extérieures",..: 3 2 1 4 6 5 7 8 9 10 ...
 $ français2: Factor w/ 26 levels "Appenzell Rhodes-Extérieures",..: 3 2 1 4 5 6 7 8 9 10 ...
 $ français3: Factor w/ 26 levels "Appenzell Rh.-E.",..: 3 2 1 4 5 6 7 8 9 10 ...
 $ allemand : Factor w/ 26 levels "Aargau","Appenzell Ausserrhoden",..: 1 3 2 6 4 5 7 8 9 10 ...
 $ allemand2: Factor w/ 26 levels "Aargau","Appenzell Ausserrhoden",..: 1 3 2 6 4 5 7 8 9 10 ...
 $ allemand3: Factor w/ 26 levels "Aargau","Appenzell A.-Rh.",..: 1 3 2 6 4 5 7 8 9 10 ...
 $ italien  : Factor w/ 26 levels "Appenzello Esterno",..: 3 2 1 6 4 5 7 8 10 11 ...
 $ english  : Factor w/ 26 levels "Appenzell Inner Rhodes",..: 3 1 2 6 5 4 7 8 9 10 ...
 $ romanche : Factor w/ 26 levels "Appenzell Dadens",..: 3 1 2 6 4 5 7 8 10 11 ...
 $ mixed    : Factor w/ 26 levels "Aargau","Appenzell Ausserrhoden",..: 1 3 2 6 4 5 7 8 9 10 ...
 $ order    : int  19 15 16 2 13 12 10 25 8 18 ...
 $ isLatin  : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ lon      : num  8.26 9.43 9.3 7.44 7.6 ...
 $ lat      : num  47.4 47.3 47.4 46.9 47.6 ...
}}
\usage{
canton_CH

canton_namesStrict(query, output = "iso2")
}
\arguments{
\item{query}{a character vector of Swiss canton names in any language}

\item{output}{a character string, the ouput format}
}
\description{
Functions to convert Switzerland's cantons, districts and communes names to their abbreviations and official BFS ID
TODO: expand countrycode for more languages (see in inst/ countrynames.csv )

Convert Swiss canton names to different languages and abbreviations
}
\details{
See \code{canton_CH}
}
\examples{
print(canton_CH)
canton_query <- c('Argovie', 'Appenzell Rh.-I.', 'St. Gallen', 'Zürich', 'randomChar', 'Vaudx')
canton_namesStrict(canton_query)
canton_namesStrict(canton_query, output = "english")
}
\keyword{datasets}

