% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilemap_Switzerland.R
\name{tilemap_ch}
\alias{tilemap_ch}
\title{Tilemap Switzerland}
\source{
\url{https://github.com/ernstbaslerpartner/Switzerland_Tilemap}
}
\usage{
tilemap_ch(withLake = F)
}
\arguments{
\item{withLake, }{a logical have Leman and Bodensee lakes?}
}
\description{
Tilemap Switzerland
}
